We have noticed an issue with Emailing Invoices for WAHVE. BCC variable is not setting and it is sending emails to previous user. Also when we try to create temp files they are not getting created because of slash with in the customer names. Few other issues addressed as well. We stopped processing invoices that are marked with status 1. Only 0 status records are processed now.
SMTP over TLS is failing because for outlook365.com AUTH is by default not enabled for plain connections. So we were not able to initiate STARTTLS. Made change to check AUTH and STATRTTLS not enabled then do not do any authentication. Other wise do one of them. 

Also for MMD identified an issue with folder count settings to 0 when a user users "Check New Mail" with in the application. With in the where clause username check is missed.